requirements.txt
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
discord.py==1.5.1
python-dotenv==0.15.0
psycopg2==2.8.6